Handle out of sync offline database
Upon logging in and also when reconnecting, if it is detected that the stored offline data is out of sync with the servers (i.e. We haven't connected for > 45 days, meaning we would most likely have missed some entity events), then we just delete the stored offline data. Directly refreshing the offline cache is not taken care of with this commit, and we may or may not decide to do that in the future